obdsyncd: removed from CVS demos/obdfsclean.sh: now stops pupd before trying to unload obdfs module class/obdcontrol: corrected packing for COPY, MIGR IOCTL methods

+++ lustre/include/linux/obd_class.h
@@ -440,11 +440,11 @@ int gen_copy_data(struct obd_conn *dst_conn, struct obdo *dst,
  *  ======== OBD IOCL Declarations ===========
  */
 struct oic_generic {
-	int  att_connid;
-	int  att_typelen;
-	void *att_type;
-	int  att_datalen;
-	void *att_data;
+	long	 att_connid;
+	int	 att_typelen;
+	void	*att_type;
+	int	 att_datalen;
+	void	*att_data;
 };
